requirements.txt
--index-url https://phonepe.mycloudrepo.io/public/repositories/phonepe-pg-sdk-python
--extra-index-url https://pypi.org/simple
phonepe_sdk==2.1.2
Django==5.1.2
python-decouple==3.8
django-ninja==1.3.0
pillow==11.0.0
django-imagekit==5.0.0
django-treebeard==4.7.1
gunicorn
whitenoise==6.8.2
psycopg2-binary==2.9.10
django-cors-headers==4.6.0
django-cleanup==9.0.0
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
django-ninja-jwt==5.3.5
google-auth==2.38.0
django_summernote==0.8.20.0
twilio==9.4.4
django-redis==5.4.0
plivo==4.59.1
openpyxl==3.1.5
django_elasticsearch_dsl==8.2
elasticsearch-dsl==8.18.0